requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.30.5
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.0.0
python-dotenv==1.0.0
pyyaml==6.0.1
markdown==3.6
pypdf==4.2.0
tqdm==4.66.4
plotly==5.19.0
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.38.2
torch==2.2.0
sentence-transformers==2.7.0
chromadb>=0.4.0
psycopg2-binary>=2.9.9
google-genai>=0.2.0